match

+++ ugv_program/state_machine/state_machine.py
@@ -34,44 +34,43 @@ class RobotStateMachine:
 
     def transition(self, event, data=None):
         print(f"Event {RobotEvent(event).name} has occurred")
-        match event:
-            case RobotEvent.START_INIT:
-                self.state_setter(RobotState.INITIAL)
-                self.init_robot()
-                self.transition(RobotEvent.FINISH_INIT)
-            case RobotEvent.FINISH_INIT:
-                self.state_setter(RobotState.IDLE)
-                self.idle_mode()
-            case RobotEvent.START_MISN:
-                self.state_setter(RobotState.AUTONOMOUS)
-                self.auto_mode()
-            case RobotEvent.HALT_MISN:
-                self.state_setter(RobotState.IDLE)
-                self.idle_mode()
-            case RobotEvent.AUTO:
-                self.state_setter(RobotState.AUTONOMOUS)
-                self.auto_mode()
-            case RobotEvent.CONTROL:
-                self.state_setter(RobotState.CONTROLLED)
-                self.control_mode()
-            case RobotEvent.UPDATE:
-                self.state_setter(RobotState.UPDATING)
-                self.update_robot(data)
-                self.transition(RobotEvent.BACK)
-            case RobotEvent.DELETE:
-                self.state_setter(RobotState.INACTIVE)
-                self.delete_robot()
-            case RobotEvent.BACK:
-                prev = self.states.back()
-                match prev:
-                    case RobotState.IDLE:
-                        self.transition(RobotEvent.HALT_MISN)
-                    case RobotState.AUTONOMOUS:
-                        self.transition(RobotEvent.AUTO)
-                    case RobotState.CONTROLLED:
-                        self.transition(RobotEvent.CONTROL)
-                    case _:
-                        print("Invalid state")
+
+        if event == RobotEvent.START_INIT:
+            self.state_setter(RobotState.INITIAL)
+            self.init_robot()
+            self.transition(RobotEvent.FINISH_INIT)
+        elif event == RobotEvent.FINISH_INIT:
+            self.state_setter(RobotState.IDLE)
+            self.idle_mode()
+        elif event == RobotEvent.START_MISN:
+            self.state_setter(RobotState.AUTONOMOUS)
+            self.auto_mode()
+        elif event == RobotEvent.HALT_MISN:
+            self.state_setter(RobotState.IDLE)
+            self.idle_mode()
+        elif event == RobotEvent.AUTO:
+            self.state_setter(RobotState.AUTONOMOUS)
+            self.auto_mode()
+        elif event == RobotEvent.CONTROL:
+            self.state_setter(RobotState.CONTROLLED)
+            self.control_mode()
+        elif event == RobotEvent.UPDATE:
+            self.state_setter(RobotState.UPDATING)
+            self.update_robot(data)
+            self.transition(RobotEvent.BACK)
+        elif event == RobotEvent.DELETE:
+            self.state_setter(RobotState.INACTIVE)
+            self.delete_robot()
+        elif event == RobotEvent.BACK:
+            prev = self.states.back()
+            if prev == RobotState.IDLE:
+                self.transition(RobotEvent.HALT_MISN)
+            elif prev == RobotState.AUTONOMOUS:
+                self.transition(RobotEvent.AUTO)
+            elif prev == RobotState.CONTROLLED:
+                self.transition(RobotEvent.CONTROL)
+            else:
+                print("Invalid state")
 
     def init_robot(self):
         skipped, data = config_interface()
@@ -167,40 +166,40 @@ class RobotStateMachine:
         if elem == None:
             return
         _, data = elem
-        match data["command"]:
-            case "start":
-                self.transition(RobotEvent.START_MISN)
-            case "pause":
-                self.transition(RobotEvent.HALT_MISN)
-            case "continue":
-                self.transition(RobotEvent.BACK)
-            case "end":
-                self.transition(RobotEvent.HALT_MISN)
-            case "update":
-                new_data = {"NAME": data["name"], "PASSWORD": data["password"]}
-                self.transition(RobotEvent.UPDATE, new_data)
-            case "delete":
-                self.transition(RobotEvent.DELETE)
-            case "auto":
-                if self.states.cur() == RobotState.CONTROLLED:
-                    self.transition(RobotEvent.AUTO)
-            case "control":
-                if self.states.cur() == RobotState.AUTONOMOUS:
-                    self.transition(RobotEvent.CONTROL)
-            case _:
-                print("Invalid admin command")
+        command = data["command"]
+        if command == "start":
+            self.transition(RobotEvent.START_MISN)
+        elif command == "pause":
+            self.transition(RobotEvent.HALT_MISN)
+        elif command == "continue":
+            self.transition(RobotEvent.BACK)
+        elif command == "end":
+            self.transition(RobotEvent.HALT_MISN)
+        elif command == "update":
+            new_data = {"NAME": data["name"], "PASSWORD": data["password"]}
+            self.transition(RobotEvent.UPDATE, new_data)
+        elif command == "delete":
+            self.transition(RobotEvent.DELETE)
+        elif command == "auto":
+            if self.states.cur() == RobotState.CONTROLLED:
+                self.transition(RobotEvent.AUTO)
+        elif command == "control":
+            if self.states.cur() == RobotState.AUTONOMOUS:
+                self.transition(RobotEvent.CONTROL)
+        else:
+            print("Invalid admin command")
 
     def check_motion_queue(self):
         elem = motion_queue.pop(0) if motion_queue else None
         if elem == None:
             return
         _, data = elem
-        match data["device"]:
-            case "motor":
-                move_motor(data["value"])
-            case "camera":
-                move_camera(data["value"])
-            case "arm":
-                move_arm(data["command"])
-            case _:
-                print("Invalid motion command")
+        dev = data["device"]
+        if dev == "motor":
+            move_motor(data["value"])
+        elif dev == "camera":
+            move_camera(data["value"])
+        elif dev == "arm":
+            move_arm(data["command"])
+        else:
+            print("Invalid motion command")
